Rename generic type
