+++ auth-data/api/current.api
@@ -274,15 +274,19 @@ package com.google.android.horologist.auth.data.tokenshare {
     property public abstract kotlinx.coroutines.flow.Flow<T> flow;
   }
 
-  @com.google.android.horologist.auth.data.ExperimentalHorologistAuthDataApi public final class TokenBundleRepositoryImpl<TokenBundle> implements com.google.android.horologist.auth.data.tokenshare.TokenBundleRepository<TokenBundle> {
-    ctor public TokenBundleRepositoryImpl(com.google.android.horologist.data.WearDataLayerRegistry registry, androidx.datastore.core.Serializer<TokenBundle> serializer, String path);
-    method public kotlinx.coroutines.flow.Flow<TokenBundle> getFlow();
-    property public kotlinx.coroutines.flow.Flow<TokenBundle> flow;
-    field public static final com.google.android.horologist.auth.data.tokenshare.TokenBundleRepositoryImpl.Companion Companion;
+}
+
+package com.google.android.horologist.auth.data.tokenshare.impl {
+
+  @com.google.android.horologist.auth.data.ExperimentalHorologistAuthDataApi public final class TokenBundleRepositoryImpl<T> implements com.google.android.horologist.auth.data.tokenshare.TokenBundleRepository<T> {
+    ctor public TokenBundleRepositoryImpl(com.google.android.horologist.data.WearDataLayerRegistry registry, androidx.datastore.core.Serializer<T> serializer, String path);
+    method public kotlinx.coroutines.flow.Flow<T> getFlow();
+    property public kotlinx.coroutines.flow.Flow<T> flow;
+    field public static final com.google.android.horologist.auth.data.tokenshare.impl.TokenBundleRepositoryImpl.Companion Companion;
   }
 
   public static final class TokenBundleRepositoryImpl.Companion {
-    method public <TokenBundle> com.google.android.horologist.auth.data.tokenshare.TokenBundleRepositoryImpl<TokenBundle> create(com.google.android.horologist.data.WearDataLayerRegistry registry, androidx.datastore.core.Serializer<TokenBundle> serializer, optional String key);
+    method public <T> com.google.android.horologist.auth.data.tokenshare.impl.TokenBundleRepositoryImpl<T> create(com.google.android.horologist.data.WearDataLayerRegistry registry, androidx.datastore.core.Serializer<T> serializer, optional String key);
   }
 
 }
